+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 * 【core   】     isMactchRegex失效标记，增加isMatchRegex（issue#I1IPJG@Gitee）
 * 【core   】     优化Validator.isChinese
 * 【core   】     ArrayUtil.addAll增加原始类型支持（issue#898@Github）
+* 【core   】     DateUtil.parse支持2020-1-1这类日期解析（issue#I1HGWW@Github）
 
 ### Bug修复
 * 【core   】     修复SimpleCache死锁问题（issue#I1HOKB@Gitee）